+++ packages/axle/src/use.ts
@@ -1,15 +1,16 @@
 import { ref, type Ref } from 'vue'
-import { type AxleRequestConfig, type RunnerMethod, type AxleInstance } from './instance.js'
 import { isFunction } from '@varlet/shared'
+import { type RunnerMethod, type AxleInstance, type AxleRequestConfig } from './instance.js'
 
-export interface RunOptions<P> {
+export interface RunOptions<V, P> {
   url?: string
   params?: P
   config?: AxleRequestConfig
   resetValue?: boolean
+  cloneResetValue?: boolean | ((value: V) => V)
 }
 
-export type Run<R, P> = (options?: RunOptions<P>) => Promise<R>
+export type Run<V, R, P> = (options?: RunOptions<V, P>) => Promise<R>
 
 export interface UseAxleRefs<V> {
   value: Ref<V>
@@ -20,13 +21,14 @@ export interface UseAxleRefs<V> {
 }
 
 export interface UseAxleOptions<V = any, R = any, P = Record<string, any>> {
-  url: string
+  url: string | (() => string)
   method: RunnerMethod
   value?: V
   params?: P | (() => P)
   resetValue?: boolean
-  config?: AxleRequestConfig
+  cloneResetValue?: boolean | ((value: V) => V)
   immediate?: boolean
+  config?: AxleRequestConfig | (() => AxleRequestConfig)
   onBefore?(refs: UseAxleRefs<V>): void
   onAfter?(refs: UseAxleRefs<V>): void
   onTransform?(response: R, refs: UseAxleRefs<V>): V
@@ -34,15 +36,20 @@ export interface UseAxleOptions<V = any, R = any, P = Record<string, any>> {
   onError?(error: Error, refs: UseAxleRefs<V>): void
 }
 
+export interface ResetValueOptions<V> {
+  cloneResetValue?: boolean | ((value: V) => V)
+}
+
 export type UseAxleInstance<V, R, P> = [
   value: Ref<V>,
-  run: Run<R, P>,
+  run: Run<V, R, P>,
   extra: {
     uploadProgress: Ref<number>
     downloadProgress: Ref<number>
     loading: Ref<boolean>
     error: Ref<Error | undefined>
     abort(): void
+    resetValue(options?: ResetValueOptions<V>): void
   }
 ]
 
@@ -52,6 +59,10 @@ export interface CreateUseAxleOptions {
   onTransform?(response: any, refs: any): any
 }
 
+export function normalizeValueGetter<T>(valueGetter: T | (() => T)) {
+  return isFunction(valueGetter) ? valueGetter() : valueGetter
+}
+
 export function createUseAxle(options: CreateUseAxleOptions) {
   const { axle, onTransform: defaultOnTransform, immediate: defaultImmediate = false } = options
 
@@ -59,13 +70,14 @@ export function createUseAxle(options: CreateUseAxleOptions) {
     options: UseAxleOptions<V, R, P>
   ): UseAxleInstance<V, R, P> => {
     const {
-      url,
+      url: initialUrlOrGetter,
       method,
       immediate = defaultImmediate,
       value: initialValue,
       resetValue: initialResetValue,
+      cloneResetValue: initialCloneResetValue,
       params: initialParamsOrGetter,
-      config: initialConfig,
+      config: initialConfigOrGetter,
       onBefore = () => {},
       onAfter = () => {},
       onTransform = (defaultOnTransform as UseAxleOptions<V, R, P>['onTransform']) ??
@@ -74,7 +86,6 @@ export function createUseAxle(options: CreateUseAxleOptions) {
       onError = () => {},
     } = options
 
-    const initialUrl = url
     const value = ref(initialValue) as Ref<V>
     const loading = ref(false)
     const error = ref<Error>()
@@ -91,44 +102,52 @@ export function createUseAxle(options: CreateUseAxleOptions) {
 
     let controller = new AbortController()
 
-    const run: Run<R, P> = async (options: RunOptions<P> = {}) => {
+    const resetValue = (options: ResetValueOptions<V> = {}) => {
+      const cloneResetValue = options.cloneResetValue ?? initialCloneResetValue ?? false
+      const cloneFn =
+        cloneResetValue === true
+          ? (v: V) => (v == null ? null : JSON.parse(JSON.stringify(v)))
+          : isFunction(initialCloneResetValue)
+          ? initialCloneResetValue
+          : (v: V) => v
+      value.value = cloneFn(initialValue as V)
+    }
+
+    const run: Run<V, R, P> = async (options: RunOptions<V, P> = {}) => {
       if (controller.signal.aborted) {
         controller = new AbortController()
       }
 
-      const resetValue = options.resetValue ?? initialResetValue ?? false
-      if (resetValue === true) {
-        value.value = initialValue as V
+      const resetValueOption = options.resetValue ?? initialResetValue ?? false
+      if (resetValueOption === true) {
+        resetValue(options)
       }
 
       uploadProgress.value = 0
       downloadProgress.value = 0
 
-      const url = options.url ?? initialUrl
+      const url = options.url ?? normalizeValueGetter(initialUrlOrGetter)
+      const params = options.params ?? normalizeValueGetter(initialParamsOrGetter)
+      const config = options.config ?? normalizeValueGetter(initialConfigOrGetter)
 
       onBefore(refs)
 
       loading.value = true
 
       try {
-        const getterParams = isFunction(initialParamsOrGetter) ? initialParamsOrGetter() : {}
-        const response = await axle[method](
-          url,
-          { ...getterParams, ...options.params },
-          {
-            signal: controller.signal,
+        const response = await axle[method](url, params, {
+          signal: controller.signal,
 
-            onUploadProgress(event) {
-              uploadProgress.value = event.progress ?? 0
-            },
+          onUploadProgress(event) {
+            uploadProgress.value = event.progress ?? 0
+          },
 
-            onDownloadProgress(event) {
-              downloadProgress.value = event.progress ?? 0
-            },
+          onDownloadProgress(event) {
+            downloadProgress.value = event.progress ?? 0
+          },
 
-            ...options.config,
-          }
-        )
+          ...config,
+        })
 
         value.value = onTransform(response as R, refs)
         error.value = undefined
@@ -153,9 +172,9 @@ export function createUseAxle(options: CreateUseAxleOptions) {
 
     if (immediate) {
       run({
-        url: initialUrl,
-        params: isFunction(initialParamsOrGetter) ? ({} as P) : initialParamsOrGetter,
-        config: initialConfig,
+        url: normalizeValueGetter(initialUrlOrGetter),
+        params: normalizeValueGetter(initialParamsOrGetter),
+        config: normalizeValueGetter(initialConfigOrGetter),
       })
     }
 
@@ -168,6 +187,7 @@ export function createUseAxle(options: CreateUseAxleOptions) {
         uploadProgress,
         downloadProgress,
         abort,
+        resetValue,
       },
     ]
   }